+++ source/common/http/conn_pool_base.h
@@ -44,7 +44,8 @@ public:
                        const Network::ConnectionSocket::OptionsSharedPtr& options,
                        const Network::TransportSocketOptionsSharedPtr& transport_socket_options,
                        Random::RandomGenerator& random_generator,
-                       Upstream::ClusterConnectivityState& state, std::vector<Http::Protocol> protocol);
+                       Upstream::ClusterConnectivityState& state,
+                       std::vector<Http::Protocol> protocol);
 
   // ConnectionPool::Instance
   void addDrainedCallback(DrainedCb cb) override { addDrainedCallbackImpl(cb); }
@@ -80,8 +81,8 @@ protected:
 // An implementation of Envoy::ConnectionPool::ActiveClient for HTTP/1.1 and HTTP/2
 class ActiveClient : public Envoy::ConnectionPool::ActiveClient {
 public:
-  ActiveClient(HttpConnPoolImplBase& parent, uint64_t lifetime_stream_limit,
-               uint64_t concurrent_stream_limit)
+  ActiveClient(HttpConnPoolImplBase& parent, uint32_t lifetime_stream_limit,
+               uint32_t concurrent_stream_limit)
       : Envoy::ConnectionPool::ActiveClient(parent, lifetime_stream_limit,
                                             concurrent_stream_limit) {
     // The static cast makes sure we call the base class host() and not
@@ -109,9 +110,7 @@ public:
   void onEvent(Network::ConnectionEvent event) override {
     parent_.onConnectionEvent(*this, codec_client_->connectionFailureReason(), event);
   }
-  size_t numActiveStreams() const override {
-    return codec_client_->numActiveRequests();
-  }
+  uint32_t numActiveStreams() const override { return codec_client_->numActiveRequests(); }
   uint64_t id() const override { return codec_client_->id(); }
 
   Http::CodecClientPtr codec_client_;
